requirements.txt
asgiref==3.7.2
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.4
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
